[docs] Add sorting to Showcase
